Refactor column PHP type

+++ tests/Type/DecimalTest.php
@@ -39,7 +39,7 @@ final class DecimalTest extends TestCase
         $tableSchema = $db->getTableSchema('decimal_default');
 
         $this->assertSame('decimal(38,0)', $tableSchema?->getColumn('Mydecimal')->getDbType());
-        $this->assertSame('double', $tableSchema?->getColumn('Mydecimal')->getPhpType());
+        $this->assertSame('float', $tableSchema?->getColumn('Mydecimal')->getPhpType());
         $this->assertSame(38, $tableSchema?->getColumn('Mydecimal')->getSize());
         $this->assertSame(9.9999999999999998e+037, $tableSchema?->getColumn('Mydecimal')->getDefaultValue());
 
@@ -87,7 +87,7 @@ final class DecimalTest extends TestCase
         $tableSchema = $db->getTableSchema('decimal_default');
 
         $this->assertSame('decimal(38,0)', $tableSchema?->getColumn('Mydecimal')->getDbType());
-        $this->assertSame('double', $tableSchema?->getColumn('Mydecimal')->getPhpType());
+        $this->assertSame('float', $tableSchema?->getColumn('Mydecimal')->getPhpType());
         $this->assertSame(38, $tableSchema?->getColumn('Mydecimal')->getSize());
         $this->assertSame(9.9999999999999998e+037, $tableSchema?->getColumn('Mydecimal')->getDefaultValue());
